Fix subtle issue with constraint addition and removal.
As previously handled, the components were fully re-rendered
whenever a constraint above them was removed. This meant that
any information that had been entered was lost to the user,
even though it was preserved in the JavaScript.

Solution required maintaining the same key for each constraint
regardless of addition or removal of other constraints, which
meant using an object, which meant using $set and $delete
global functions.